Boxer compat localtest781

+++ src/cpu/core_full/op.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2002-2020  The DOSBox Team
+ *  Copyright (C) 2002-2021  The DOSBox Team
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -417,7 +417,7 @@ switch (inst.code.op) {
 			CPU_VERW(inst_op1_d);
 			goto nextopcode;		/* Else value will saved */
 		default:
-			LOG(LOG_CPU,LOG_ERROR)("Group 6 Illegal subfunction %X",inst.rm_index);
+			LOG(LOG_CPU,LOG_ERROR)("Group 6 Illegal subfunction %X", static_cast<uint32_t>(inst.rm_index));
 			goto illegalopcode;
 		}
 		break;
@@ -453,7 +453,7 @@ switch (inst.code.op) {
 			PAGING_ClearTLB();
 			goto nextopcode;
 		default:
-			LOG(LOG_CPU,LOG_ERROR)("Group 7 Illegal subfunction %X",inst.rm_index);
+			LOG(LOG_CPU,LOG_ERROR)("Group 7 Illegal subfunction %X", static_cast<uint32_t>(inst.rm_index));
 			goto illegalopcode;
 		}
 		break;
@@ -662,6 +662,6 @@ switch (inst.code.op) {
 	case 0:
 		break;
 	default:
-		LOG(LOG_CPU,LOG_ERROR)("OP:Unhandled code %d entry %X",inst.code.op,inst.entry);
+		LOG(LOG_CPU,LOG_ERROR)("OP:Unhandled code %d entry %X",inst.code.op,static_cast<uint32_t>(inst.entry));
 		
 }
